requirements.txt
attrs==16.0.0
cffi==1.7.0
cryptography==1.4
cssselect==0.9.2
dateparser==0.4.0
enum34==1.1.6
feedparser==5.2.1
idna==2.1
ipaddress==1.0.16
jdatetime==1.8.1
lxml==3.6.0
parsel==1.0.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
python-dateutil==2.5.3
pytz==2016.4
queuelib==1.4.2
regex==2016.6.19
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.11
Scrapy==1.1.0
service-identity==16.0.0
six==1.10.0
Twisted==16.2.0
umalqurra==0.2
w3lib==1.14.2
zope.interface==4.2.0
